requirements.txt
blinker==1.8.2
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Mail==0.10.0
Flask-SSLify==0.1.5
Flask-WTF==1.2.1
gunicorn==23.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.1
pillow==10.4.0
python-dotenv==1.0.1
reportlab==4.2.2
requests==2.32.3
stripe==10.7.0
typing_extensions==4.12.2
urllib3==2.2.2
Werkzeug==3.0.3
WTForms==3.1.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
psycopg2-binary==2.9.6